Sessions gets a multicast client.
To achieve this, we move the multicast middleware completely out of the client middleware, making it top-level in buster-server.

The session middleware gets one single multicast client exposed to it. The multicast ID and URL is exposed when creating sessions, and is the same id and URL for every session.