Changing the way we create buster-multicast clients.
We don't pre-create a multicast middleware client with a captured browser. Instead, the client is created when the page loads, so that reloading a client page is the same as visiting it the first time. We use an identifier - the id of the captured browser - and events to attach the multicast middleware client to the captured browser.