Testing serving of slave page and header.
Using a buster-resources middleware to host the slave resource set and header resource set.

Using the main slave resource set for the hosting of the browser capture page as well, instead of having a separate respond method for that. This causes the context path we serve this resource set from to be /slaves/123/browser, to avoid conflicts with /slaves/123/header.