More queue-like, less event-like.
We use one /slave_ready event for listening to slaves becoming
ready. Along with the slave id we transmit the faye client id, which we
will later use for listening to heartbeats and know when a slave dies.

Using one event lets us not have to manage as many subscriptions, and it
better suits the upcoming message queue semantics for knowing when a
slave dies.